+++ libraries/session/src/main/java/androidx/media3/session/MediaLibrarySessionImpl.java
@@ -256,12 +256,10 @@ import java.util.concurrent.Future;
 
   public void notifyChildrenChanged(
       String parentId, int itemCount, @Nullable LibraryParams params) {
-    dispatchRemoteControllerTaskWithoutReturn(
-        (callback, seq) -> {
-          if (isSubscribed(callback, parentId)) {
-            callback.onChildrenChanged(seq, parentId, itemCount, params);
-          }
-        });
+    List<ControllerInfo> connectedControllers = instance.getConnectedControllers();
+    for (int i = 0; i < connectedControllers.size(); i++) {
+      notifyChildrenChanged(connectedControllers.get(i), parentId, itemCount, params);
+    }
   }
 
   public void notifyChildrenChanged(